Create new team and player definitions
Note:
The only function associated with these classes is the removeCard
function in the player class. This function is now put with the
match tools, and it is changed to just take any hand and remove
a card from that hand. It will likely be updated later to return
the hand and the card object that was removed.